implement merge_and_save_mfdataset
Fixes #244
