+++ pyopia/pipeline.py
@@ -7,11 +7,14 @@ from typing import TypedDict
 import datetime
 import pandas as pd
 from operator import methodcaller
-import toml
 import sys
 import logging
+import pyopia.io
 
 logger = logging.getLogger()
+
+
+steps_from_xstats = pyopia.io.steps_from_xstats
 
 
 class Pipeline():
@@ -271,23 +274,6 @@ def steps_to_string(steps):
                       + '\n   Vars: ' + str(vars(steps[key]))
                       + '\n')
     return steps_str
-
-
-def steps_from_xstats(xstats):
-    '''Get the steps attribute from xarray version of the particle stats into a dictionary
-
-    Parameters
-    ----------
-    xstats : xarray.DataSet
-        xarray version of the particle stats dataframe, containing metadata
-
-    Returns
-    -------
-    dict
-        TOML-formatted dictionary of pipeline steps
-    '''
-    steps = toml.loads(xstats.__getattr__('steps'))
-    return steps
 
 
 def build_repr(toml_steps, step_name):
